apps/producer/requirements.txt
kafka-python-ng==2.2.3
requests==2.32.5
python-dotenv==1.1.1